(original rev 2150) 64 bit (just incase)
git-svn-id: file:///tmp/current/omni/branches/staging@1910 2ffc8166-1744-45da-a82a-d6ea10f7896b

+++ omni/src/volume/omMipVolume.cpp
@@ -1461,7 +1461,7 @@ void OmMipVolume::AllocMemMapFiles()
 		}
 		file->resize(size);
 		
-		for( qint64 i=0; i < size; i+=4096){
+		for( qint64 i=0; i < size; i+=(qint64)4096){
 			file->seek(i);
 			file->putChar(0);
 		}
